requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
psycopg2-binary